+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 [Unreleased]## [0.5.8.1b0]
+### Added
+* ScenarioDbTable._set_df_column_types added support for Time type for time-of-day columns
+* DataManager.get_parameter_value added support for Time type for time-of-day parameters
+### Fixed
 
 ## [0.5.8.0]- 2026-01-23
 ### Fixed